requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
bs4==0.0.2
cachelib==0.9.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
fasttext==0.9.3
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gevent==24.2.1
gevent-websocket==0.10.1
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-core==2.4.1
google-cloud-datastore==2.20.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.11
langchain-core==0.2.26
langchain-text-splitters==0.2.2
langsmith==0.1.95
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
mysqlclient==2.2.4
numpy==1.24.4
orjson==3.10.6
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.1
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
requests==2.32.3
rq==1.16.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
soupsieve==2.5
SQLAlchemy==2.0.31
tenacity==8.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
zope.event==5.0
zope.interface==6.4.post2
